Title: 달에서 무게 구하기, Time: 95ms, Memory: 9MB, Status: Passed - Codetree

+++ 260204/달에서 무게 구하기/weight-on-the-moon.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
         // Please write your code here.
         int w = 13;
-        int g = 0.165;
+        double g = 0.165;
         System.out.printf("%d * %.6f = %.6f", w, g, w*g);
     }
 }
